EXC-981: Update provisional create to support 128-bit cycles

This MR updates provisional create args to support 128-bit cycles. 

See merge request dfinity-lab/public/ic!3530

+++ rs/replica_tests/tests/tests.rs
@@ -17,7 +17,7 @@ use ic_types::{
 };
 
 const WASM_PAGE_SIZE: usize = 65536;
-const CYCLES_BALANCE: u64 = 1 << 50;
+const CYCLES_BALANCE: u128 = 1 << 50;
 
 #[test]
 /// Tests a message can roundtrip through all layers
